connecting to backend

+++ backend/src/main/java/org/example/project/issueController.java
@@ -84,18 +84,24 @@ public class issueController {
             throw new IllegalArgumentException("issueId must not be null");
         }
 
+        if (token == null || !token.startsWith("Bearer ")) {
+            throw new IllegalArgumentException("Invalid Authorization header format.");
+        }
+
+        token = token.substring(7); // Extract the token part
         String email = jwtutils.extractEmailFromToken(token);
         if (email == null) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
 
         User user = userService.findUserByEmail(email);
-        issueService.deleteIssue(issueId,user.getId());
+        issueService.deleteIssue(issueId, user.getId());
         MessageResponse res = new MessageResponse();
         res.setMessage("Issue deleted");
 
         return ResponseEntity.ok(res);
     }
+
 
     @PutMapping("/{issueId}/assignee/{userId}")
     public ResponseEntity<Issue> addUserToIssue(@PathVariable Long issueId,
